better destroy logic. wip

+++ src/stages/StageOne.js
@@ -9,7 +9,8 @@ export default class StageOne {
   camera
   scene
   renderer
-  hittableComponents = [] //{componentId:'box123', threejsObject: new THREE.Mesh( geometry, material)}
+  hittableComponents = [] //[component, component]
+  hitteeComponents = []//[component, component]
   rendererDomElement //created during initThreejs and used by react component to append to appropriate div
   children = []
   raycaster = new Raycaster()//used for hit tests
@@ -36,40 +37,6 @@ export default class StageOne {
 
     this.rendererDomElement = this.renderer.domElement;
     animate({camera, scene, renderer:this.renderer, stage:this});
-  }
-
-  hitTestV3({hittableComponents=this.hittableComponents, camera=this.camera, projector=this.projector, clientX, clientY}){
-    // let {width, height} = this.getScreenDimensions();
-    // let mouseX = (clientX / width) * 2 - 1;
-    // let mouseY = - (clientY / height) * 2 + 1;
-    // let mouseVector = new Vector3(mouseX, mouseY, 1);
-    //
-    // projector.unprojectVector(mouseVector, camera);
-    //
-    // var direction = mouseVector.sub(camera.position).normalize();
-    // var ray = new Raycaster(camera.position, direction);
-    //
-    // let hitComponent;//first object hit by ray
-    // let intersects;//raycaster intersectObject result
-    // let threejsObjects = hittableComponents.map(h=>h.threejsObject);
-    // let allIntersects = ray.intersectObjects(threejsObjects, true);
-    //
-    // for(let i of allIntersects){
-    //   let clickedThreejsObject = i.object;
-    //   for(let h of hittableComponents){
-    //     if(h.threejsObject.name == clickedThreejsObject.name){
-    //       hitComponent = h;
-    //       intersects = [i];
-    //       break;
-    //     }
-    //   }
-    //   if(hitComponent){
-    //     break;
-    //   }
-    // }
-    //
-    // if(hitComponent == undefined){return;}
-    // return {hitComponent, intersects};
   }
 
   fireBullet({camera=this.camera, scene=this.scene, projector=this.projector, clientX, clientY}){
@@ -131,20 +98,7 @@ export default class StageOne {
     //all registered hittable components will be evaluated to determine if the mouse x, y coordinates intersect/hit.
     //https://threejs.org/docs/#api/core/Raycaster
     [ec.mouse.mousedown]({clientX, clientY, raycaster=this.raycaster}){
-      // console.log('checking for hit components...');
       this.fireBullet({clientX, clientY});
-      //
-      // let {hitComponent, intersects} = this.hitTestV3({clientX, clientY}) || {};
-      // if(!hitComponent){return;}
-      //
-      // console.log(`hit component: ${hitComponent.componentId}`);
-      // let {scene} = this;
-      // intersects.forEach(i=>{
-      //   let {distance, point, face, faceIndex, indices, object} = i;
-      //   let {componentId} = hitComponent;
-      //   signal.trigger(ec.hitTest.hitComponent, {componentId, distance, point, face, faceIndex, indices, object, scene});
-      // })
-
     },
     //anything that wants to be hittable (e.g. by a bullet) should register via this signal
     [ec.hitTest.registerHittableComponent]({component}){
@@ -157,13 +111,16 @@ export default class StageOne {
       if(hitIndex < 0){return;}
       this.hittableComponents.splice(hitIndex, 1);//remove hittable component from array
     },
-    [ec.stage.componentDestroyed]({componentId, threejsObject}){
-      signal.trigger(ec.hitTest.unregisterHittableComponent, {componentId});
-    },
+
+    //bullet calls this when its finished its distance and after it hits something.
     [ec.stage.destroyComponent]({componentId, scene=this.scene}){
-      let component = this.removeChild({componentId});
-      signal.trigger(ec.hitTest.unregisterHittableComponent, {componentId});
-      component.destroy({scene});
+      //since this is done while iterating over children, don't modify the array while looping. wait until next tick
+      setTimeout(function(){
+        let component = this.removeChild({componentId});
+        if(!component){return;} //same componentId over and over.
+        component.destroy({scene});
+      }.bind(this), 0)
+
     }
   }
